requirements.txt
Flask
gunicorn
python-dotenv
azure-communication-identity